+++ development.nix
@@ -2,8 +2,7 @@
 # Development environment tools and configurations. #
 #####################################################
 {...}: {
-  # The `systems` attribute is a list of supported systems used by the
-  # `perSystem` attribute to define system-specific attributes of the flake.
+  # Define the systems where the development environment is available.
   systems = [
     "aarch64-linux"
     "x86_64-linux"
@@ -11,9 +10,7 @@
     "x86_64-darwin"
   ];
 
-  # The `perSystem` attribute defines system-specific attributes of the flake,
-  # making the following tools and configurations available on all systems
-  # defined in the `systems` attribute.
+  # Define the development environment configurations.
   perSystem = {
     config,
     pkgs,
@@ -49,16 +46,15 @@
     devshells.default = {
       # Configure the shell information.
       devshell = {
-        name = "dotfiles";
+        name = "dotfiles-dev";
         motd = ''
-          {2}{bold}[i] Development shell for Mario's dotfiles repository.
+          {2}{bold}[i] Development shell for Mario's dotfiles repository.\
           {reset}
-          {3}[!] Don't use commands from the internal category.{reset}
           $(type -p menu &> /dev/null && menu)
         '';
         startup."setup-env".text = ''
           # Install pre-commit hooks.
-          _pre-commit-install
+          ${config.pre-commit.installationScript}
         '';
       };
 
@@ -68,14 +64,6 @@
         {
           package = pkgs.just;
           help = "Type `just` to see the available recipes.";
-        }
-
-        # Add an internal command to install pre-commit hooks.
-        {
-          name = "_pre-commit-install";
-          command = config.pre-commit.installationScript;
-          help = "Internal command to install pre-commit hooks.";
-          category = "internal";
         }
       ];
     };