+++ home-configurations/mariovagomarzal/core.nix
@@ -1,7 +1,23 @@
 #######################
 # Core home packages. #
 #######################
-{pkgs, ...}: {
+{
+  pkgs,
+  config,
+  ...
+}: {
+  # Let home-manager install and manage itself.
+  programs.home-manager.enable = true;
+
+  # User information for home-manager.
+  home = {
+    homeDirectory = "/Users/${config.home.username}";
+
+    # Supported version.
+    stateVersion = "23.11";
+  };
+
+  # No-configuration packages.
   home.packages = with pkgs; [
     # Files related tools.
     fzf